I built a retention model and identified factors driving attrition.
There are 50000 agents at my org and there is high attrition among them. It is a very serious issue as it is hard to find a high number of replacements within a short time. This also affects the KPIs.  I built a retention model and identified factors driving attrition. Also suggested recommendations to prevent attrition.
I built a Logistic regression and random forest on this data. I selected logistic regression as there was no performance difference between them. One of the challenges I faced during this was to increase the model accuracy above a threshold accepted by the organization. Since our goal is to reduce false negatives, as the cost of False negatives is more than False positives.